some fixes

+++ src/main/java/org/example/orissemwork/services/FileService.java
@@ -1,6 +1,6 @@
 package org.example.orissemwork.services;
 
-import org.example.orissemwork.db.FileInfoDAO;
+import org.example.orissemwork.dao.FileInfoDAO;
 import org.example.orissemwork.model.FileInfo;
 import org.example.orissemwork.model.User;
 
@@ -18,28 +18,36 @@ public class FileService {
         this.fileInfoDAO = fileInfoDAO;
     }
 
-    public void saveFileToStorage(InputStream stream, String type, Long size, String name, User user) throws IOException, SQLException {
+    private static String pathToFolder = "files/";
 
+    public void saveFileToStorage(InputStream stream, String type, Long size, String name, User user) throws IOException, SQLException {
         FileInfo file = new FileInfo(null, name, UUID.randomUUID().toString(), size, type);
         String path = "/home/fewalthel/IdeaProjects/ORIS-semwork/src/main/webapp/files/" + file.getStorageFileName() + "." + file.getType().split("/")[1];
 
         Files.copy(stream, Paths.get(path));
         fileInfoDAO.save(file, user);
-
-        System.out.println("загрузка файла");
     }
 
 
     public String getPathToAvatarForUser(User user) throws SQLException {
         FileInfo fileInfo = fileInfoDAO.findByIdUser(user.getId());
+
         if (fileInfo == null) {
-            return "files/default_avatar.jpg";
+            return pathToFolder + "default_avatar.jpeg";
         } else {
-            return "files/" + fileInfo.getStorageFileName() + "." + fileInfo.getType().split("/")[1];
+            return pathToFolder + fileInfo.getStorageFileName() + "." + fileInfo.getType().split("/")[1];
         }
     }
 
-    public void deleteAvatarFromStorage(User user) throws SQLException {
-        fileInfoDAO.deleteFromDB(user);
+
+    public void setAvatar(InputStream stream, String type, Long size, String name, User user) throws SQLException, IOException {
+        FileInfo fileInfo = fileInfoDAO.findByIdUser(user.getId());
+
+        if (fileInfo != null) {
+            fileInfoDAO.deleteFromDB(user);
+            /*String path = pathToFolder + fileInfo.getStorageFileName() + "." + fileInfo.getType().split("/")[1];
+            Files.delete(Paths.get(path));*/
+        }
+        saveFileToStorage(stream, type, size, name, user);
     }
 }
